pyproject.toml
[project]
name = "skvaider"
version = "0.1.0dev0"
license = "MIT"
description = "Router for ML/AI model API access"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Flying Circus Internet Operations GmbH", email = "mail@flyingcircus.io" },
]
dependencies = [
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "greenlet>=3.2.4",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "svcs>=25.1.0",
    "uvicorn[standard]>=0.35.0",
    "rfc8785>=0.1.4",
    "structlog>=25.4.0",
]

[project.scripts]
skvaider = "skvaider:main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
line_length = 80

[dependency-groups]
dev = [
    "openai>=1.100.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-timeout>=2.4.0",
]
